# MyMusicBoxApi/http/play.go
package http

import (
	"fmt"
	"musicboxapi/configuration"
	"musicboxapi/logging"
	"musicboxapi/models"
	"net/http"
	"os"

	"github.com/gin-gonic/gin"
)

// @Produce json
// @Param sourceId path string true "Id of song/file to serve using http 206 partial content"
// @Description Enables playback for song/file using http 206 partial content
// @Success 200 "serve song/file with range request (http 206)"
// @Failure 500 {object} models.ApiResponseModel
// @Router /api/v1/play/:sourceId [get]
func Play(ctx *gin.Context) {
	sourceId := ctx.Param("sourceId")

	if sourceId == "" {
		ctx.Status(http.StatusInternalServerError)
		logging.Warning("No sourceId in request")
		return
	}

	path := fmt.Sprintf("%s/%s.%s", configuration.Config.SourceFolder, sourceId, configuration.Config.OutputExtension)

	file, err := os.Open(path)

	if err != nil {
		logging.Error(fmt.Sprintf("Failed to read file, %s", err.Error()))
		ctx.JSON(http.StatusInternalServerError, models.ErrorResponse(err))
		return
	}
	defer file.Close()

	fileInfo, err := file.Stat()
	if err != nil {
		logging.Error(fmt.Sprintf("Failed could not get fileinfo: %s", err.Error()))
		ctx.JSON(http.StatusInternalServerError, models.ErrorResponse(err))
		return
	}

	// Set proper headers for OPUS audio
	ctx.Header("Content-Type", "audio/opus")
	ctx.Header("Accept-Ranges", "bytes")

	// Use http.ServeContent which handles range requests automatically
	http.ServeContent(
		ctx.Writer,
		ctx.Request,
		fileInfo.Name(),
		fileInfo.ModTime(),
		file,
	)
}